requirements.txt
asgiref==3.7.2
Django==5.0.2
django-phonenumber-field==7.3.0
mysqlclient==2.2.4
phonenumbers==8.13.30
pillow==10.2.0
pipfile==0.0.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tzdata==2022.2
